Update seo_url.php
"I noticed that in OpenCart 4.x, when I entered a page name in the URL that doesn't exist, the system didn't display a 'Page Not Found' error as it used to. Instead, it showed a different kind of error. I believe it would be beneficial to reinstate the 'Page Not Found' error message for better user experience and clear navigation."